When the user trys to delete a member who is part of a training, show a error message. #22

+++ tvwTeilnehmerliste-frontend/src/app/management/management.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Trainer } from '../trainer';
 import { Participant } from '../participant';
 import { TvwApiService } from '../tvw-api.service';
+import { Training } from 'app/training';
+import { log } from 'util';
 
 @Component({
   selector: 'app-management',
@@ -12,6 +14,10 @@ export class ManagementComponent implements OnInit {
 
   trainers: Trainer[];
   participants: Participant[];
+  trainings: Training[];
+
+  showDeleteTrainerError = 'hideError';
+  showDeleteParticipantError = 'hideError';
 
 
   constructor(
@@ -21,6 +27,7 @@ export class ManagementComponent implements OnInit {
   ngOnInit() {
     this.getTrainers();
     this.getParticipants();
+    this.getTrainings();
   }
 
   getTrainers(): void {
@@ -33,7 +40,16 @@ export class ManagementComponent implements OnInit {
       .then(participants => this.participants = participants);
   }
 
+  getTrainings(): void {
+    this.tvwApiService.getTrainings()
+      .then(trainings => this.trainings = trainings);
+  }
+
   deleteTrainer(id: number): void {
+    if(this.hasMemberATraining(id, true)){
+      this.showDeleteTrainerError = 'showError';
+      return;
+    }
     this.tvwApiService.deleteTrainer(id)
       .then(() => {
         this.trainers = this.trainers.filter(trainer => trainer.id !== id);
@@ -41,11 +57,31 @@ export class ManagementComponent implements OnInit {
   }
 
   deleteParticipant(id: number): void {
+    if(this.hasMemberATraining(id,false)){
+      this.showDeleteParticipantError = 'showError';
+      return;
+    }
     this.tvwApiService.deleteParticipant(id)
       .then(() => {
         this.participants = this.participants.filter(participant => participant.id !== id);
       });
   }
 
+  private hasMemberATraining(memberId: number, isTrainer): boolean {
+    if(isTrainer){
+        for (let training of this.trainings) {
+            if(training.trainers.find(x => x.id == memberId)){
+              return true;
+            }
+        }
+    }else {
+        for (let training of this.trainings) {
+          if(training.participants.find(x => x.id == memberId)){
+            return true;
+          }
+        }
+    }
+    return false;
+}
 
 }
